Put log parsers in module
